requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
algoliasearch==2.6.2
amqp==2.6.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto==2.49.0
boto3==1.24.36
botocore==1.27.36
celery==4.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-appconf==1.0.5
django-celery==3.3.1
django-compressor==4.0
django-countries==7.2
django-countries-with-calling-codes==0.1.1
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-flexible-subscriptions==0.15.1
django-htmlmin==0.11.0
django-model-utils==4.2.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-randompinfield==0.4.7
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-s3-cache==1.4.3
django-storages==1.12.3
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
frozenlist==1.3.0
geoip2==4.5.0
html5lib==1.1
idna==3.3
jmespath==1.0.1
kombu==4.6.11
maxminddb==2.2.0
multidict==6.0.2
oauthlib==3.2.0
paypalrestsdk==1.13.1
phonenumbers==8.12.50
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.1
rcssmin==1.1.0
requests==2.28.0
requests-oauthlib==1.3.1
rjsmin==1.2.0
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
twilio==7.9.2
urllib3==1.26.9
user-profile==1.0.0
vine==1.3.0
webencodings==0.5.1
yarl==1.7.2
